+++ plugin/fileLoader/mod.ts
@@ -27,3 +27,5 @@ export function pluginFileLoader(opts: Opts = {}): Plugin {
     },
   };
 }
+
+export default pluginFileLoader;