+++ plugin/mod.ts
@@ -1,7 +1,6 @@
 export * from "./hooks.ts";
 export * from "./importResolver/mod.ts";
 export * from "./rootResolver/mod.ts";
-export * from "./cacheResolver/mod.ts";
 export * from "./chainResolver/mod.ts";
 export { pluginCacheLoader, Opts as CacheLoaderOptions } from "./cacheLoader/mod.ts";
 export { pluginFileLoader, Opts as FileLoaderOptions } from "./fileLoader/mod.ts";