+++ plugin/typescriptTransform/mod.ts
@@ -9,3 +9,5 @@ export function pluginTypescriptTransform(opts?: Deno.CompilerOptions): Plugin {
     },
   };
 }
+
+export default pluginTypescriptTransform;