+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Plugins for html and css assets
 - React template recipe
+- Plugins are now also exported with `export default`
 
 ### Changed
 
 - Updated Rollup to v2.26.3
 - Toplevel mod.ts now re-exports rollup from deps.ts
+- Removed pluginCacheResolver
 
 ## [0.5.0](https://github.com/denofn/denopack/compare/0.4.0..0.5.0) - 2020-08-13